requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-filter==2.3.0
djangorestframework==3.11.1
idna==2.10
importlib-metadata==1.7.0
Markdown==3.2.2
psycopg2-binary==2.8.5
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
zipp==3.1.0
